Portfolio/IBKR_Trading_Engine/ib_client/requirements.txt
selenium==4.27.1
requests==2.32.3
slack-sdk==3.34.0
twilio==9.4.2
webdriver-manager
python-dotenv
